+++ ItunesMedia/Library/Extensions/Codable+asDictionary.swift
@@ -1,0 +1,19 @@
+//
+//  Codable+toDictionary.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/26/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import Foundation
+
+extension Encodable {
+    func asDictionary() throws -> [String: Any] {
+        let data = try JSONEncoder().encode(self)
+        guard let dictionary = try JSONSerialization.jsonObject(with: data, options: .allowFragments) as? [String: Any] else {
+            throw NSError()
+        }
+        return dictionary
+    }
+}