+++ ItunesMedia/Models/ResponseModels/ResponseMediaModel.swift
@@ -1,0 +1,16 @@
+//
+//  ResponseMediaModel.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/26/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import Foundation
+
+struct ResponseMediaModel: Media, Decodable {
+    let trackId: Int
+    let artistName: String
+    let artworkUrl60: String
+    let releaseDate: String
+}